+++ cmd/sointu-vsti/nameid_not_native.go
@@ -1,0 +1,6 @@
+//go:build !native
+
+package main
+
+var PLUGIN_ID = [4]byte{'S', 'n', 't', 'u'}
+var PLUGIN_NAME = "Sointu"